+++ src/Gallifrey.Jira/Model/WorkLogs.cs
@@ -5,7 +5,7 @@ namespace Gallifrey.Jira.Model
     public class WorkLogs
     {
         public List<WorkLog> worklogs { get; set; }
-        public int total { get; set; }
-        public int maxResults { get; set; }
+        public double total { get; set; }
+        public double maxResults { get; set; }
     }
 }
